+++ src/com/shopbilling/services/BillingServices.java
@@ -127,7 +127,7 @@ public class BillingServices {
 				stmt = conn.prepareStatement(UPDATE_PRODUCT_STOCK);
 				for(ItemDetails item : itemList){
 					stmt.setInt(2, item.getItemNo());
-					stmt.setInt(1, item.getQuantity());
+					stmt.setDouble(1, item.getQuantity());
 					stmt.addBatch();
 				}
 				int batch[] = stmt.executeBatch();
