+++ src/com/shopbilling/services/SalesReturnServices.java
@@ -52,7 +52,7 @@ public class SalesReturnServices {
 				stmt.setDate(6,bill.getBillDate());
 				stmt.setString(7, bill.getBillPaymentMode());
 				stmt.setInt(8, bill.getNoOfItems());
-				stmt.setInt(9, bill.getTotalQuanity());
+				stmt.setDouble(9, bill.getTotalQuanity());
 				stmt.setString(10, bill.getReturnpaymentMode());
 				stmt.setDouble(11, bill.getTotalAmount());
 				stmt.setDouble(12, bill.getBillNetSalesAmt());
@@ -93,7 +93,7 @@ public class SalesReturnServices {
 				stmt = conn.prepareStatement(UPDATE_PRODUCT_STOCK);
 				for(ItemDetails item : itemList){
 					stmt.setInt(2, item.getItemNo());
-					stmt.setInt(1, item.getQuantity());
+					stmt.setDouble(1, item.getQuantity());
 					stmt.addBatch();
 				}
 				int batch[] = stmt.executeBatch();
@@ -128,7 +128,7 @@ public class SalesReturnServices {
 					stmt.setInt(2, item.getItemNo());
 					stmt.setDouble(3,item.getMRP());
 					stmt.setDouble(4, item.getRate());
-					stmt.setInt(5, item.getQuantity());
+					stmt.setDouble(5, item.getQuantity());
 					stmt.setDouble(6, item.getAmount());
 					stmt.addBatch();
 				}
@@ -171,7 +171,7 @@ public class SalesReturnServices {
 					returnDetails.setCustomerMobileNo(rs.getLong("CUST_MOB_NO"));
 					returnDetails.setCustomerName(rs.getString("CUSTOMER_NAME"));
 					returnDetails.setNoOfItems(rs.getInt("NO_OF_ITEMS"));
-					returnDetails.setTotalQuanity(rs.getInt("TOTAL_QTY"));
+					returnDetails.setTotalQuanity(rs.getDouble("TOTAL_QTY"));
 					returnDetails.setTotalAmount(rs.getDouble("RETURN_TOTAL_AMOUNT"));
 					returnDetails.setReturnpaymentMode(rs.getString("PAYMENT_MODE"));
 					returnDetails.setNewBillnetSalesAmt(rs.getDouble("NEW_BILL_NET_SALES_AMOUNT"));
@@ -213,7 +213,7 @@ public class SalesReturnServices {
 					returnDetails.setCustomerMobileNo(rs.getLong("CUST_MOB_NO"));
 					returnDetails.setCustomerName(rs.getString("CUSTOMER_NAME"));
 					returnDetails.setNoOfItems(rs.getInt("NO_OF_ITEMS"));
-					returnDetails.setTotalQuanity(rs.getInt("TOTAL_QTY"));
+					returnDetails.setTotalQuanity(rs.getDouble("TOTAL_QTY"));
 					returnDetails.setTotalAmount(rs.getDouble("RETURN_TOTAL_AMOUNT"));
 					returnDetails.setReturnpaymentMode(rs.getString("PAYMENT_MODE"));
 					returnDetails.setNewBillnetSalesAmt(rs.getDouble("NEW_BILL_NET_SALES_AMOUNT"));
@@ -247,7 +247,7 @@ public class SalesReturnServices {
 				stmt = conn.prepareStatement(UPDATE_PRODUCT_STOCK);
 				for(ItemDetails item : itemList){
 					stmt.setInt(2, item.getItemNo());
-					stmt.setInt(1, item.getQuantity());
+					stmt.setDouble(1, item.getQuantity());
 					stmt.addBatch();
 				}
 				int batch[] = stmt.executeBatch();
@@ -311,7 +311,7 @@ public class SalesReturnServices {
 				itemDetails.setItemName(rs.getString("PRODUCT_NAME"));
 				itemDetails.setMRP(rs.getDouble("ITEM_MRP"));
 				itemDetails.setRate(rs.getDouble("ITEM_RATE"));
-				itemDetails.setQuantity(rs.getInt("ITEM_QTY"));
+				itemDetails.setQuantity(rs.getDouble("ITEM_QTY"));
 				
 				itemDetailsList.add(itemDetails);
 			}
@@ -362,7 +362,7 @@ public class SalesReturnServices {
 					returnDetails.setCustomerMobileNo(rs.getLong("CUST_MOB_NO"));
 					returnDetails.setCustomerName(rs.getString("CUSTOMER_NAME"));
 					returnDetails.setNoOfItems(rs.getInt("NO_OF_ITEMS"));
-					returnDetails.setTotalQuanity(rs.getInt("TOTAL_QTY"));
+					returnDetails.setTotalQuanity(rs.getDouble("TOTAL_QTY"));
 					returnDetails.setTotalAmount(rs.getDouble("RETURN_TOTAL_AMOUNT"));
 					returnDetails.setReturnpaymentMode(rs.getString("PAYMENT_MODE"));
 					returnDetails.setNewBillnetSalesAmt(rs.getDouble("NEW_BILL_NET_SALES_AMOUNT"));
